server/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
pymongo[tls,srv]==3.12.0
python-dotenv==0.19.0
Gunicorn
Werkzeug==2.0.2